# specs/openapi.yaml
openapi: 3.0.3
info:
  title: Jobswyft API
  version: 1.0.0
  description: AI-powered job application assistant API
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.jobswyft.com
    description: Production
paths:
...
  /v1/jobs:
    post:
      summary: Create a new job record
      description: |
        Creates a new job record with optional status. Use this endpoint when manually saving
        a job as "applied" or with a specific status. Defaults to "saved" if no status provided.
      operationId: createJob
      tags:
        - Jobs
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
            examples:
              withAppliedStatus:
                summary: Save job as applied
                value:
                  title: "Senior Software Engineer"
                  company: "Acme Corp"
                  description: "We are looking for..."
                  source_url: "https://acme.com/jobs/123"
                  status: "applied"
              withDefaultStatus:
                summary: Save job (default status)
                value:
                  title: "Backend Developer"
                  company: "TechCo"
                  description: "Join our team..."
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          description: Validation error (missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List tracked jobs with pagination
      description: |
        Returns a paginated list of tracked jobs with optional filtering by status.
        Includes a notes preview (first 100 characters) for each job.
      operationId: listJobs
      tags:
        - Jobs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/JobStatus'
          description: Filter by job status
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
        - in: query
          name: sort
          schema:
            type: string
            default: updated_at
          description: Field to sort by
      responses:
        '200':
          description: Paginated list of jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
              example:
                success: true
                data:
                  items:
                    - id: "550e8400-e29b-41d4-a716-446655440000"
                      title: "Senior Software Engineer"
                      company: "Acme Corp"
                      status: "applied"
                      notes_preview: "Recruiter mentioned..."
                      created_at: "2026-01-30T12:00:00Z"
                      updated_at: "2026-01-30T14:00:00Z"
                  total: 25
                  page: 1
                  page_size: 20
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/jobs/scan:
    post:
      summary: Save scanned job from extension
      operationId: createScannedJob
      tags:
        - Jobs
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          description: Validation error (missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/jobs/{job_id}:
    get:
      summary: Get job details
      operationId: getJob
      tags:
        - Jobs
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: job_id
          schema:
            type: string
            format: uuid
          required: true
          description: Job UUID
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update job fields
      operationId: updateJob
      tags:
        - Jobs
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: job_id
          schema:
            type: string
            format: uuid
          required: true
          description: Job UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobUpdateRequest'
      responses:
        '200':
          description: Job updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a job
      operationId: deleteJob
      tags:
        - Jobs
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: job_id
          schema:
            type: string
            format: uuid
          required: true
          description: Job UUID
      responses:
        '200':
          description: Job deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDeleteResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/jobs/{job_id}/status:
    put:
      summary: Update job status
      description: |
        Updates only the job status field. Valid statuses are:
        saved, applied, interviewing, offered, rejected, accepted
      operationId: updateJobStatus
      tags:
        - Jobs
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: job_id
          schema:
            type: string
            format: uuid
          required: true
          description: Job UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobStatusUpdateRequest'
            example:
              status: "interviewing"
      responses:
        '200':
          description: Job status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Invalid status. Allowed: saved, applied, interviewing, offered, rejected, accepted"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/jobs/{job_id}/notes:
    put:
      summary: Update job notes
      description: |
        Adds or updates notes on a job. Notes are replaced entirely (not appended).
      operationId: updateJobNotes
      tags:
        - Jobs
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: job_id
          schema:
            type: string
            format: uuid
          required: true
          description: Job UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobNotesUpdateRequest'
            example:
              notes: "Recruiter mentioned team is scaling. Follow up next week."
      responses:
        '200':
          description: Job notes updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: AUTH_REQUIRED
            message:
              type: string
            details:
              type: object
...
    JobStatus:
      type: string
      enum:
        - saved
        - applied
        - interviewing
        - offered
        - rejected
        - accepted
      description: Job application status
    JobCreateRequest:
      type: object
      required:
        - title
        - company
        - description
      properties:
        title:
          type: string
          example: Senior Software Engineer
        company:
          type: string
          example: Acme Corp
        description:
          type: string
          example: We are looking for a senior engineer...
        location:
          type: string
          nullable: true
          example: San Francisco, CA
        salary_range:
          type: string
          nullable: true
          example: $150k-$200k
        employment_type:
          type: string
          nullable: true
          example: Full-time
        source_url:
          type: string
          nullable: true
          format: uri
          example: https://acme.com/jobs/123
        status:
          $ref: '#/components/schemas/JobStatus'
          description: Optional status (defaults to "saved" if not provided)
    JobUpdateRequest:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        description:
          type: string
        location:
          type: string
          nullable: true
        salary_range:
          type: string
          nullable: true
        employment_type:
          type: string
          nullable: true
        source_url:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/JobStatus'
        notes:
          type: string
          nullable: true
    JobData:
      type: object
      required:
        - id
        - user_id
        - title
        - company
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
          example: Senior Software Engineer
        company:
          type: string
          example: Acme Corp
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        salary_range:
          type: string
          nullable: true
        employment_type:
          type: string
          nullable: true
        source_url:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/JobStatus'
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JobResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/JobData'
    JobListItem:
      type: object
      required:
        - id
        - title
        - company
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Senior Software Engineer
        company:
          type: string
          example: Acme Corp
        status:
          $ref: '#/components/schemas/JobStatus'
        notes_preview:
          type: string
          nullable: true
          description: First 100 characters of notes (if any)
          example: "Recruiter mentioned..."
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JobListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - items
                - total
                - page
                - page_size
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/JobListItem'
                total:
                  type: integer
                  description: Total number of matching jobs
                  example: 25
                page:
                  type: integer
                  description: Current page number
                  example: 1
                page_size:
                  type: integer
                  description: Items per page
                  example: 20
    JobStatusUpdateRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/JobStatus'
    JobNotesUpdateRequest:
      type: object
      required:
        - notes
      properties:
        notes:
          type: string
          description: Notes content to save
          example: "Recruiter mentioned team is scaling. Follow up next week."
    JobDeleteResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
                  example: Job deleted successfully
...